% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolation_function}
\alias{interpolation_function}
\title{Create an interpolation function}
\usage{
interpolation_function(x, y, type)
}
\arguments{
\item{x}{Independent variable}

\item{y}{Dependent variable}

\item{type}{Character string indicating the interpolation type
("constant", "linear" or "spline").}
}
\description{
Create an interpolation function
}
